Refactor: update ActionBar, ContentView; add TextFormatter, KeyHandler
