+++ Sources/HarryFanReader/SettingsView.swift
@@ -16,30 +16,8 @@ struct SettingsView: View {
     // Dismiss environment value for closing the view
     @Environment(\.dismiss) private var dismiss
 
-    // Currently selected font
-    @State private var selectedFont: FontManager.MSDOSFont = .vdu8x16
-    // Current font size
-    @State private var fontSize: Double = 16.0
-    // Remove excessive empty lines toggle
-    @State private var removeEmptyLines: Bool = true
-    // Word wrap toggle
-    @State private var wordWrap: Bool = true
-    // Wrap width value
-    @State private var wrapWidth: Double = 80.0
-    // Original font for cancel functionality
-    @State private var originalFont: FontManager.MSDOSFont = .vdu8x16
-    // Original font size for cancel functionality
-    @State private var originalFontSize: Double = 16.0
-    // Original remove empty lines value for cancel functionality
-    @State private var originalRemoveEmptyLines: Bool = true
-    // Original word wrap value for cancel functionality
-    @State private var originalWordWrap: Bool = true
-    // Original wrap width value for cancel functionality
-    @State private var originalWrapWidth: Double = 80.0
-    // Anti-aliasing toggle
-    @State private var enableAntiAliasing: Bool = true
-    // Original anti-aliasing value for cancel functionality
-    @State private var originalEnableAntiAliasing: Bool = true
+    // View model for settings
+    @StateObject private var viewModel = SettingsViewModel(fontManager: FontManager(), document: TextDocument())
 
     // Main view body rendering the settings UI
     var body: some View {
@@ -59,9 +37,6 @@ struct SettingsView: View {
                         .onTapGesture {
                             dismiss()
                         }
-                        .onHover { _ in
-                            // Hover effect would go here
-                        }
 
                     Circle()
                         .fill(Color.yellow)
@@ -71,12 +46,6 @@ struct SettingsView: View {
                                 .stroke(Color.black.opacity(0.2), lineWidth: 0.5),
                         )
                         .scaleEffect(1.0)
-                        .onTapGesture {
-                            // Minimize functionality
-                        }
-                        .onHover { _ in
-                            // Hover effect would go here
-                        }
 
                     Circle()
                         .fill(Color.green)
@@ -86,12 +55,6 @@ struct SettingsView: View {
                                 .stroke(Color.black.opacity(0.2), lineWidth: 0.5),
                         )
                         .scaleEffect(1.0)
-                        .onTapGesture {
-                            // Maximize functionality
-                        }
-                        .onHover { _ in
-                            // Hover effect would go here
-                        }
                 }
                 .padding(.leading, 12)
 
@@ -131,7 +94,7 @@ struct SettingsView: View {
             // Settings content
             Form {
                 Section("Font Settings") {
-                    Picker("Font", selection: $selectedFont) {
+                    Picker("Font", selection: $viewModel.selectedFont) {
                         ForEach(FontManager.MSDOSFont.allCases, id: \.self) { font in
                             Text(font.displayName).tag(font)
                         }
@@ -139,24 +102,24 @@ struct SettingsView: View {
 
                     HStack {
                         Text("Font Size")
-                        Slider(value: $fontSize, in: 8 ... 32, step: 1)
-                        Text("\(Int(fontSize))")
+                        Slider(value: $viewModel.fontSize, in: 8 ... 32, step: 1)
+                        Text("\(Int(viewModel.fontSize))")
                             .frame(width: 30)
                     }
                 }
 
                 Section("Text Processing") {
-                    Toggle("Remove excessive empty lines", isOn: $removeEmptyLines)
+                    Toggle("Remove excessive empty lines", isOn: $viewModel.removeEmptyLines)
                         .help("Automatically remove multiple consecutive empty lines")
 
-                    Toggle("Word wrap", isOn: $wordWrap)
+                    Toggle("Word wrap", isOn: $viewModel.wordWrap)
                         .help("Automatically wrap long lines to fit the display width")
 
-                    if wordWrap {
+                    if viewModel.wordWrap {
                         HStack {
                             Text("Wrap width")
-                            Slider(value: $wrapWidth, in: 40 ... 200, step: 1)
-                            Text("\(Int(wrapWidth))")
+                            Slider(value: $viewModel.wrapWidth, in: 40 ... 200, step: 1)
+                            Text("\(Int(viewModel.wrapWidth))")
                                 .frame(width: 40)
                         }
                         .help("Maximum number of characters per line")
@@ -164,7 +127,7 @@ struct SettingsView: View {
                 }
 
                 Section("Display") {
-                    Toggle("Enable anti-aliasing", isOn: $enableAntiAliasing)
+                    Toggle("Enable anti-aliasing", isOn: $viewModel.enableAntiAliasing)
                         .help("Smooth text rendering for better visual quality")
                 }
             }
@@ -176,38 +139,17 @@ struct SettingsView: View {
 
                 // Cancel button restores original values and dismisses
                 Button("Cancel") {
-                    // Restore original values
-                    selectedFont = originalFont
-                    fontSize = originalFontSize
-                    removeEmptyLines = originalRemoveEmptyLines
-                    wordWrap = originalWordWrap
-                    wrapWidth = originalWrapWidth
-                    enableAntiAliasing = originalEnableAntiAliasing
+                    viewModel.cancelChanges()
                     dismiss()
                 }
                 .keyboardShortcut(.escape)
 
                 // Submit button applies changes and dismisses
                 Button("Submit") {
-                    // Apply changes
-                    fontManager.currentFont = selectedFont
-                    fontManager.fontSize = fontSize
-                    document.removeEmptyLines = removeEmptyLines
-                    document.wordWrap = wordWrap
-                    document.wrapWidth = Int(wrapWidth)
-
-                    // Apply anti-aliasing setting
-                    AppSettings.enableAntiAliasing = enableAntiAliasing
-
-                    // Reload file with new settings if a file is open
-                    if !document.fileName.isEmpty {
-                        document.reloadWithNewSettings()
-                    }
-
+                    viewModel.applySettings(fontManager: fontManager, document: document)
                     dismiss()
                 }
                 .keyboardShortcut(.return)
-                .buttonStyle(.borderedProminent)
             }
             .padding()
             .background(Color(NSColor.controlBackgroundColor))
@@ -218,25 +160,16 @@ struct SettingsView: View {
                 alignment: .top,
             )
         }
-        .frame(width: 450, height: 400)
+        .frame(width: 400, height: 500)
         .background(Color(NSColor.windowBackgroundColor))
-        // onAppear stores original values for cancel functionality and sets current values
         .onAppear {
-            // Store original values for cancel functionality
-            originalFont = fontManager.currentFont
-            originalFontSize = fontManager.fontSize
-            originalRemoveEmptyLines = document.removeEmptyLines
-            originalWordWrap = document.wordWrap
-            originalWrapWidth = Double(document.wrapWidth)
-            originalEnableAntiAliasing = AppSettings.enableAntiAliasing
-
-            // Set current values
-            selectedFont = fontManager.currentFont
-            fontSize = fontManager.fontSize
-            removeEmptyLines = document.removeEmptyLines
-            wordWrap = document.wordWrap
-            wrapWidth = Double(document.wrapWidth)
-            enableAntiAliasing = AppSettings.enableAntiAliasing
+            // Initialize view model with current values
+            viewModel.selectedFont = fontManager.currentFont
+            viewModel.fontSize = fontManager.fontSize
+            viewModel.removeEmptyLines = document.removeEmptyLines
+            viewModel.wordWrap = document.wordWrap
+            viewModel.wrapWidth = Double(document.wrapWidth)
+            viewModel.enableAntiAliasing = AppSettings.enableAntiAliasing
         }
     }
 }
